refactor: move utility function `pad_element_bboxes()` to pdf_image_utils.py
